requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
numpy
opencv-python-headless
streamlit
Pillow
ultralytics
pytest